+++ doh-server/server.go
@@ -48,6 +48,7 @@ type Server struct {
 type DNSRequest struct {
 	request         *dns.Msg
 	response        *dns.Msg
+	transactionID   uint16
 	currentUpstream string
 	isTailored      bool
 	errcode         int
@@ -159,10 +160,15 @@ func (s *Server) handlerFunc(w http.ResponseWriter, r *http.Request) {
 		jsonDNS.FormatError(w, fmt.Sprintf("Invalid argument value: \"ct\" = %q", contentType), 415)
 		return
 	}
+	if req.errcode == 444 {
+		return
+	}
 	if req.errcode != 0 {
 		jsonDNS.FormatError(w, req.errtext, req.errcode)
 		return
 	}
+
+	req = s.patchRootRD(req)
 
 	var err error
 	req, err = s.doDNSQuery(req)
@@ -207,6 +213,16 @@ func (s *Server) findClientIP(r *http.Request) net.IP {
 		return ip
 	}
 	return nil
+}
+
+// Workaround a bug causing Unbound to refuse returning anything about the root
+func (s *Server) patchRootRD(req *DNSRequest) *DNSRequest {
+	for _, question := range req.request.Question {
+		if question.Name == "." {
+			req.request.RecursionDesired = true
+		}
+	}
+	return req
 }
 
 func (s *Server) doDNSQuery(req *DNSRequest) (resp *DNSRequest, err error) {